requirements.txt
click
environs==7.4.0
python-dotenv
environs
azureml-pipeline-steps
azureml-pipeline-core
azureml-sdk
azureml-pipeline
azureml-core
marshmallow==3.26.2
